+++ services/transcription-service/src/gc/gc-storage.service.ts
