+++ services/transcription-service/src/transcription/transcription.module.ts
@@ -3,7 +3,6 @@ import { TranscriptionService } from './transcription.service';
 import { TranscriptionResolver } from './transcription.resolver';
 import { UsersResolver } from './user.resolver';
 import { GcModule } from 'src/gc/gc.module';
-import { PrismaService } from 'src/prisma/prisma.service';
 import { PrismaModule } from 'src/prisma/prisma.module';
 
 @Module({
