+++ services/transcription-service/src/transcription/transcription.resolver.ts
@@ -1,4 +1,4 @@
-import { Args, Int, Query, Resolver } from '@nestjs/graphql';
+import { Args, Int, Mutation, Query, Resolver } from '@nestjs/graphql';
 import { Transcription } from './transcription.model';
 import { TranscriptionService } from './transcription.service';
 
@@ -11,13 +11,13 @@ export class TranscriptionResolver {
     return this.transcriptionService.findOneById(id);
   }
 
-  @Query((returns) => String)
+  @Mutation((returns) => String)
   createSignedUrl(@Args('file_name') file_name: string) {
     const url = this.transcriptionService.createSignedUrl(file_name);
     return url;
   }
 
-  @Query((returns) => String)
+  @Mutation((returns) => String)
   createTranscription(@Args('file_name') file_name: string) {
     const url = this.transcriptionService.createTranscription(file_name);
     return url;
