+++ supergraph.graphql
@@ -3,6 +3,7 @@ schema
   @link(url: "https://specs.apollo.dev/join/v0.3", for: EXECUTION)
 {
   query: Query
+  mutation: Mutation
 }
 
 directive @join__enumValue(graph: join__Graph!) repeatable on ENUM_VALUE
@@ -40,13 +41,18 @@ enum link__Purpose {
   EXECUTION
 }
 
+type Mutation
+  @join__type(graph: TRANSCRIPTIONS)
+{
+  createSignedUrl(file_name: String!): String!
+  createTranscription(file_name: String!): String!
+}
+
 type Query
   @join__type(graph: TRANSCRIPTIONS)
   @join__type(graph: USERS)
 {
   transcription(id: Int!): Transcription! @join__field(graph: TRANSCRIPTIONS)
-  createSignedUrl(file_name: String!): String! @join__field(graph: TRANSCRIPTIONS)
-  createTranscription(file_name: String!): String! @join__field(graph: TRANSCRIPTIONS)
   fetchTranscription(transcription_id: String!): String! @join__field(graph: TRANSCRIPTIONS)
   user(id: Int!): User! @join__field(graph: USERS)
 }
